move HTTPMultiBin to it's own package
This is done because it gets in the way of using the `testutils` package
inside lib
