+++ js/runner_test.go
@@ -45,7 +45,7 @@ import (
 	"github.com/loadimpact/k6/js/modules/k6/ws"
 	"github.com/loadimpact/k6/lib"
 	"github.com/loadimpact/k6/lib/metrics"
-	"github.com/loadimpact/k6/lib/testutils"
+	"github.com/loadimpact/k6/lib/testutils/httpmultibin"
 	"github.com/loadimpact/k6/lib/types"
 	"github.com/loadimpact/k6/stats"
 	"github.com/loadimpact/k6/stats/dummy"
@@ -211,7 +211,7 @@ func TestOptionsPropagationToScript(t *testing.T) {
 }
 
 func TestMetricName(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	script := tb.Replacer.Replace(`
@@ -229,7 +229,7 @@ func TestMetricName(t *testing.T) {
 }
 
 func TestSetupDataIsolation(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	script := tb.Replacer.Replace(`
@@ -759,7 +759,7 @@ func TestVUIntegrationBlacklistScript(t *testing.T) {
 }
 
 func TestVUIntegrationHosts(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	r1, err := getSimpleRunner("/script.js",
@@ -940,7 +940,7 @@ func TestVUIntegrationOpenFunctionError(t *testing.T) {
 
 func TestVUIntegrationCookiesReset(t *testing.T) {
 
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	r1, err := getSimpleRunner("/script.js", tb.Replacer.Replace(`
@@ -990,7 +990,7 @@ func TestVUIntegrationCookiesReset(t *testing.T) {
 }
 
 func TestVUIntegrationCookiesNoReset(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	r1, err := getSimpleRunner("/script.js", tb.Replacer.Replace(`
@@ -1217,7 +1217,7 @@ func TestVUIntegrationClientCerts(t *testing.T) {
 }
 
 func TestHTTPRequestInInitContext(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	_, err := getSimpleRunner("/script.js", tb.Replacer.Replace(`
@@ -1296,7 +1296,7 @@ func TestInitContextForbidden(t *testing.T) {
 			k6metrics.ErrMetricsAddInInitContext.Error(),
 		},
 	}
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	for _, test := range table {
@@ -1314,7 +1314,7 @@ func TestInitContextForbidden(t *testing.T) {
 }
 
 func TestArchiveRunningIntegraty(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	fs := afero.NewMemMapFs()
@@ -1358,7 +1358,7 @@ func TestArchiveRunningIntegraty(t *testing.T) {
 }
 
 func TestArchiveNotPanicking(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	fs := afero.NewMemMapFs()
@@ -1380,7 +1380,7 @@ func TestArchiveNotPanicking(t *testing.T) {
 }
 
 func TestStuffNotPanicking(t *testing.T) {
-	tb := testutils.NewHTTPMultiBin(t)
+	tb := httpmultibin.NewHTTPMultiBin(t)
 	defer tb.Cleanup()
 
 	r, err := getSimpleRunner("/script.js", tb.Replacer.Replace(`
